fix bags with session and sales view

+++ src/main/java/com/example/login/controller/Projects.java
@@ -2,6 +2,7 @@ package com.example.login.controller;
 
 import java.util.ArrayList;
 
+import com.example.login.LoginUI;
 import org.hibernate.Session;
 import org.hibernate.criterion.Restrictions;
 
@@ -11,23 +12,38 @@ import com.example.login.util.HibernateUtil;
 public class Projects {
 
 	public static ArrayList<Project> getProjects(){
-		 Session session = HibernateUtil.getSessionFactory().getCurrentSession();
-		 session.beginTransaction();
+        Session session = LoginUI.getCurrentSession();
+        if(session.getTransaction() == null){
+            session.beginTransaction();
+        }
+        else
+            session.getTransaction().begin();
+
 		 ArrayList<Project> list = (ArrayList<Project>) session.createCriteria(Project.class).list();
 		 session.getTransaction().commit();	 
 		 return list;
 	 }
 	
 	public static void addPosition(Project project){
-		 Session session = HibernateUtil.getSessionFactory().getCurrentSession();
-		 session.beginTransaction();		 
+        Session session = LoginUI.getCurrentSession();
+        if(session.getTransaction() == null){
+            session.beginTransaction();
+        }
+        else
+            session.getTransaction().begin();
+
 		 session.save(project);		 
 		 session.getTransaction().commit();	 
 	 }
 	
 	public static ArrayList<Project> getByName(String name){
-		 Session session = HibernateUtil.getSessionFactory().getCurrentSession();
-		 session.beginTransaction();		 
+        Session session = LoginUI.getCurrentSession();
+        if(session.getTransaction() == null){
+            session.beginTransaction();
+        }
+        else
+            session.getTransaction().begin();
+
 		 ArrayList<Project> list = (ArrayList<Project>) session.createCriteria(Project.class).
 				 											   add(Restrictions.eq("name", name.trim())).list();		 
 		 session.getTransaction().commit();	 
@@ -35,8 +51,12 @@ public class Projects {
 	 }
 	
 	public static void delete(Project project){
-		 Session session = HibernateUtil.getSessionFactory().getCurrentSession();
-		 session.beginTransaction();
+        Session session = LoginUI.getCurrentSession();
+        if(session.getTransaction() == null){
+            session.beginTransaction();
+        }
+        else
+            session.getTransaction().begin();
 		 session.delete(project);
 		 session.getTransaction().commit();	 
 	}
